backend/requirements.txt
fastapi
uvicorn
langchain
langchain-groq
langchain-chroma
langchain-community
langchain-text-splitters
langchain-core
chromadb
sentence-transformers
pypdf
python-docx
docx2txt
python-dotenv
python-multipart
requests
watchdog
